requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dateutil
requests
python-dotenv
google-generativeai